requirements.txt
asgiref==3.6.0
attrs==22.1.0
boto3==1.26.50
botocore==1.29.50
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
Django==3.2.15
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.11
h11==0.14.0
hvac==1.0.2
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
Pillow==9.2.0
pycparser==2.21
pyhcl==0.4.4
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.11
validators==0.20.0
wsproto==1.1.0
zipp==3.11.0